Change the namespace of the Presenter class
