--- app/helpers/spree/retail/presenter_helper.rb
@@ -1,13 +1,0 @@
-module Spree
-  module Retail
-    module PresenterHelper
-      def present(object, presenter_type)
-        presenter_klass = "#{presenter_type}_presenter".camelcase
-        klass = "Shopify::#{presenter_klass}".constantize
-        presenter = klass.new(object)
-        yield presenter if block_given?
-        presenter
-      end
-    end
-  end
-end
