+++ app/helpers/spree/retail/shopify/presenter_helper.rb
@@ -1,0 +1,15 @@
+module Spree
+  module Retail
+    module Shopify
+      module PresenterHelper
+        def present(object, presenter_type)
+          presenter_klass = "#{presenter_type}_presenter".camelcase
+          klass = "Shopify::#{presenter_klass}".constantize
+          presenter = klass.new(object)
+          yield presenter if block_given?
+          presenter
+        end
+      end
+    end
+  end
+end
